requirements.txt
notebook==7.5.5
pandas<3
plotly==6.6.0
scikit-learn==1.8.0
mlflow>=2.21
gradio>=5.2
pydantic>=2.10
pyyaml>=6.0
joblib>=1.4
shap>=0.47
ruff>=0.11
pytest>=8.3
